pull lucid into a Types header
