+++ CxxTest/include/CxxTest.hpp
@@ -13,19 +13,7 @@
 #include <memory>
 
 #include "Vector.h"
-
-namespace lucid {
-
-    template<typename PrecisionT>
-    struct InstantiateTypes
-    {
-        typedef math::Vector<PrecisionT, 3> Vector3;
-        typedef math::Vector<PrecisionT, 4> Vector4;
-    };
-    
-    typedef InstantiateTypes<float> tile;
-
-}
+#include "Types.h"
 
 using Vector3 = lucid::tile::Vector3;
